add EventHistory subscriber that stores statistics in database
EventHistory subscriber listens for keypresses and stores their number
per day in an SQLite database.

```
sqlite> select * from statistics;
Date       | KeypressCnt
2025-10-21 | 87
2025-10-22 | 19
2025-10-23 | 3
```

Sqlite is probably overkill for a two-column table (I could just use csv
file), but I wanted to practice with sql and rust bindings.

One interesting case happened: after pressing "Ctrl-C", the program
closed correctly at first glance, but the "history.db-journal" file
always remained. It turned out that after SIGINT, the listener was still
waiting for the last "C" press, and the EventHistory callback function
was called, which tried to UPDATE the database, but at that point it
quit and sqlite did not delete the journal file.

After I added a SIGINT handler that stopped the listener first (other
components stopped because `tx` channel was closed), UPDATE operation
completed correctly, and the last "C" keypress was saved.

> "A rollback journal is a temporary file used to implement atomic commit
and rollback capabilities in SQLite. The rollback journal is usually
created when a transaction is first started and is usually deleted when
a transaction commits or rolls back."